simplify the example. enhance readme to be more tutorial based.
