+++ docs/sample-apps/jfk-files-mcp-server/.env.example
@@ -1,0 +1,1 @@
+MISTRAL_API_KEY=your-mistral-api-key
