+++ docs/sample-apps/jfk-files-mcp-server/Dockerfile
