+++ docs/sample-apps/jfk-files-mcp-server/requirements.txt
@@ -9,4 +9,3 @@ mcp
 httpx
 uvicorn
 starlette
-python-dotenv
